// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  candidate: 'candidate',
  recruiter: 'recruiter',
  admin: 'admin'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const OfferStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived'
} as const

export type OfferStatus = (typeof OfferStatus)[keyof typeof OfferStatus]


export const ApplicationStatus = {
  pending: 'pending',
  interview: 'interview',
  accepted: 'accepted',
  rejected: 'rejected',
  withdrawn: 'withdrawn'
} as const

export type ApplicationStatus = (typeof ApplicationStatus)[keyof typeof ApplicationStatus]


export const ParseStatus = {
  pending: 'pending',
  processing: 'processing',
  done: 'done',
  failed: 'failed'
} as const

export type ParseStatus = (typeof ParseStatus)[keyof typeof ParseStatus]


export const ImportedFrom = {
  upload: 'upload',
  json: 'json',
  github: 'github'
} as const

export type ImportedFrom = (typeof ImportedFrom)[keyof typeof ImportedFrom]